[breaking change] disable setspatialref!(geom, spatialref)
we encourage users to keep the geometries separate from the spatial reference systems for now.

(i) it is often that we have a single spatialref for an entire set of geometries, so the spatialref should be set on the geomfielddefn instead of the geometry.

(ii) this will also avoid a large class of potential memory issues.